feat: 장바구니 일반제품 추가 기능

+++ server/src/main/java/com/buyte/member/dto/CartResDto.java
@@ -11,5 +11,5 @@ public class CartResDto {
     private Long productId;
     private String productName;
     private String productImagePath;
-    private Long productPrice;
+    private Integer productPrice;
 }
